Migrate existing TF configuration to integrate with new repo
Migrate to the new repo to integrate with CI/CD.
The existing configuration consists of a main VPC with 3 subnets - 1
public and 2 private.
The public subnet will hold a load balancer which will point to the 2
web servers in the private subnet.
The other private subnet will hold the database for the web servers.
The terraform configuration uses a preprovisioned S3 bucket as a backend
for the state.