Provision EC2 instances implementing high availability
TF configuration for the 2 EC2 instances that will serve as the
Wordpress web servers.
Decided to introduce higher availability by provisioning an additional
private and public subnet(total 2 of each) which are mapped to each
other with their availability zones. This sets up the scene for the LB
implementation which will be linked to the public subnets and point to
the EC2 instances residing in the private subnets.