Provision ELB in front of the two web servers
Configure an ELB that routes the traffic to the 2 private wordpress web
servers.
Also provisioned a security group linked to the ELB that permits
internet traffic to the ELB and outbound traffic to the web servers.
Changed the SG rules for the web server SG to accommodate the load
balancer rules.