Change to ALB instead of ELB
Having too many issues settuping ELB(it is looking for a default
subnet). Since there isn't an actual need for an ELB in the current
setup an ALB seems more appropriate.
Also finally decided to test the idempotency of the configuration and
ran into an issue with the automatic AZ choice, so I had to resort to a
data resource and hard indexes for the AZs. I see this as a valid
solution.
Up to this point a full destroy and reapply is working correctly.