Add user data script to the web servers
The script installs the necessary dependencies needed for wordpress.
After that the script downloads wordpress and the wp-cli which enables
it to automate the wordpress installation and the initial database
configuration.
This user data script receives the database credentials and RDS endpoint
through the terraform configuration which does not expose anything
sensitive to the outside world.