Add missing private subnet route table
The route table for the private subnets was missing which resulted in
packet loss.
Also fixed an issue with --allow-root for the wp-cli tool in the user
data script. When running wp-cli as root(user data in AWS runs as root)
you must supply --allow-root to the commands, otherwise it returns an
error/warning and the command does not complete.